requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
curlify==2.2.1
DateTime==5.5
distconfig3==1.0.1
environs==14.1.0
Faker==35.2.0
idna==3.10
iniconfig==2.0.0
marshmallow==3.26.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyHamcrest==2.1.0
pyTelegramBotAPI==4.26.0
pytest==8.3.4
pytest-telegram-notifier==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rootpath==0.1.0
setuptools==75.8.0
six==1.17.0
structlog==25.1.0
swagger-coverage @ https://github.com/dirolorbit/swagger-coverage-py/archive/refs/heads/master.zip#sha256=c2f2548ae37091f92fe25aa9c3cf2f917fcb6303e38e1960f11fb0b386b8cc12
telebot==0.0.5
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
vyper-config==1.2.1
watchdog==6.0.0
zope.interface==7.2
